move general logic to BackgroundEvaluationStep
