+++ completion-plugin/src/org/jb/cce/evaluation/step/ActionsInterpretationStep.kt
@@ -1,43 +1,18 @@
 package org.jb.cce.evaluation.step
 
-import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.progress.ProgressManager
-import com.intellij.openapi.progress.Task
-import com.intellij.openapi.progress.impl.BackgroundableProcessIndicator
 import com.intellij.openapi.project.Project
 import com.intellij.util.concurrency.FutureResult
 import org.jb.cce.EvaluationWorkspace
 import org.jb.cce.evaluation.ActionsInterpretationHandler
+import org.jb.cce.util.Progress
 
 class ActionsInterpretationStep(project: Project, isHeadless: Boolean): BackgroundEvaluationStep(project, isHeadless) {
     override val name: String = "Actions interpreting"
 
     override val description: String = "Interpretation of generated actions"
 
-    override fun start(workspace: EvaluationWorkspace): EvaluationWorkspace? {
-        val result = FutureResult<EvaluationWorkspace?>()
-        val task = object : Task.Backgroundable(project, name) {
-
-            override fun run(indicator: ProgressIndicator) {
-                indicator.text = this.title
-                ActionsInterpretationHandler(project).invoke(workspace, workspace, getProgress(indicator))
-            }
-
-            override fun onSuccess() {
-                result.set(workspace)
-            }
-
-            override fun onCancel() {
-                evaluationAbortedHandler.onCancel(this.title)
-                result.set(null)
-            }
-
-            override fun onThrowable(error: Throwable) {
-                evaluationAbortedHandler.onError(error, this.title)
-                result.set(null)
-            }
-        }
-        ProgressManager.getInstance().runProcessWithProgressAsynchronously(task, BackgroundableProcessIndicator(task))
-        return result.get()
+    override fun evaluateStep(workspace: EvaluationWorkspace, result: FutureResult<EvaluationWorkspace?>, progress: Progress) {
+        ActionsInterpretationHandler(project).invoke(workspace, workspace, progress)
+        result.set(workspace)
     }
 }
